+++ heylogs-ext-gitlab/src/main/java/nbbrd/heylogs/ext/gitlab/GitLabSupport.java
@@ -23,15 +23,12 @@ class GitLabSupport {
         // Utility class, no instances allowed
     }
 
-    static boolean isKnownHost(@NonNull ForgeLink expected) {
-        return Arrays.asList(expected.getBase().getHost().split("\\.", -1)).contains("gitlab");
-    }
-
     // https://docs.gitlab.com/user/reserved_names/#rules-for-usernames-project-and-group-names-and-slugs
     // FIXME: This regex is not perfect, it allows some invalid names.
     static final Pattern NAMESPACE_PATTERN = Pattern.compile("[a-z\\d](?:[a-z\\d]|-(?=[a-z\\d])){0,38}", Pattern.CASE_INSENSITIVE);
     static final Pattern PROJECT_PATTERN = Pattern.compile("[a-z\\d._-]{1,100}", Pattern.CASE_INSENSITIVE);
     static final Pattern HASH_PATTERN = Pattern.compile("[0-9a-f]{7,40}", Pattern.CASE_INSENSITIVE);
+    static final Pattern OID_PATTERN = Pattern.compile(".+\\.{3}.+", Pattern.CASE_INSENSITIVE);
     static final Pattern NUMBER_PATTERN = Pattern.compile("\\d+", Pattern.CASE_INSENSITIVE);
     static final String DASH_KEYWORD = "-";
     static final String PATH_SEPARATOR = "/";